+++ Source/WindAttenuation.h
@@ -7,7 +7,7 @@ double dGetLowPassFilterCutoff(double dWindSpeed, double dDistance)
     // Compute attenuation for upwind
     double A = 0.1 * (dWindSpeed / 30.0);
     double dAttenuationdB = A * dDistance;
-    double dCutoffFreq = 20000.0 * expf(-dAttenuationdB / 10.0);
+    double dCutoffFreq = 20000.0 * exp(-dAttenuationdB / 10.0);
     dCutoffFreq = juce::jmax(50.0, dCutoffFreq);
     return dCutoffFreq;
 }
